move fallback-errors components to common folder
