+++ bigbluebutton-html5/imports/ui/components/common/fallback-errors/fallback-view/component.jsx
